fix(lint): resolve remaining gofmt, unconvert issues
- gofmt: reformat hpa_test.go, monitoring.go
- unconvert: remove redundant int32 casts in networkpolicies_test.go

+++ internal/resources/networkpolicies_test.go
@@ -304,9 +304,9 @@ var _ = Describe("APIFrontend NetworkPolicy", func() {
 				}
 			}
 		}
-		Expect(ingressPorts).To(HaveKey(int32(PortHTTPS)))
-		Expect(ingressPorts).To(HaveKey(int32(PortHealthProbe)))
-		Expect(ingressPorts).To(HaveKey(int32(PortMetrics)))
+		Expect(ingressPorts).To(HaveKey(PortHTTPS))
+		Expect(ingressPorts).To(HaveKey(PortHealthProbe))
+		Expect(ingressPorts).To(HaveKey(PortMetrics))
 	})
 
 	It("allows egress to DNS, API server, and kubernaut pods", func() {
